+++ app/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -5,10 +5,10 @@ import java.util.Map;
 import java.util.UUID;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UnauthorizedException;
-import nextstep.jwp.http.HttpRequest;
-import nextstep.jwp.http.HttpResponse;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.response.HttpResponse;
 import nextstep.jwp.http.HttpSession;
-import nextstep.jwp.http.HttpStatus;
+import nextstep.jwp.http.response.HttpStatus;
 import nextstep.jwp.model.User;
 import nextstep.jwp.utils.ContentType;
 import nextstep.jwp.utils.FileReader;
