+++ app/src/main/java/nextstep/jwp/http/RequestHandler.java
@@ -7,6 +7,9 @@ import java.net.Socket;
 import java.util.Objects;
 import nextstep.jwp.controller.Controller;
 import nextstep.jwp.controller.ExceptionHandler;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.response.HttpResponse;
+import nextstep.jwp.http.response.HttpStatus;
 import nextstep.jwp.utils.ContentType;
 import nextstep.jwp.utils.FileReader;
 import nextstep.jwp.utils.HttpRequestReader;
