+++ app/src/main/java/nextstep/jwp/http/request/HttpMethod.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.http.request;
 
 public enum HttpMethod {
