+++ app/src/main/java/nextstep/jwp/http/response/HttpStatus.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.http.response;
 
 public enum HttpStatus {
